requirements.txt
fastapi[standard]>=0.113.0,<0.114.0
pydantic>=2.7.0,<3.0.0
python-dotenv==1.0.1
psycopg==3.2.3
psycopg2-binary==2.9.9
pydantic==2.8.2
aiohttp==3.9.5
asyncpg==0.29.0
sqlmodel==0.0.22
fastapi-users[sqlalchemy,oauth]